+++ plone/formwidget/masterselect/tests/test_intergration.py
@@ -1,38 +1,23 @@
-import unittest
-from niteoweb.windmill import WindmillTestCase
-from Products.PloneTestCase.setup import setupPloneSite
-from Products.PloneTestCase.layer import onsetup
-from Products.Five.zcml import load_config
+# -*- coding: utf-8 -*-
 from Testing import ZopeTestCase as ztc
-from plone.dexterity.utils import createContent
+from plone.formwidget.masterselect.testing import \
+    PLONE_FORMWIDGET_MASTERSELECT_INTEGRATION
+import unittest
 
 
-@onsetup
-def load_zcml():
-    import plone.app.dexterity
-    load_config('configure.zcml', plone.app.dexterity)
-    ztc.installPackage('plone.app.dexterity')
+class DecoUITestCase(unittest.TestCase):
 
-load_zcml()
-#setupPloneSite(products=['plone.app.page'])
-setupPloneSite(products=['plone.app.dexterity'])
+    layer = PLONE_FORMWIDGET_MASTERSELECT_INTEGRATION
 
-
-class DecoUITestCase(WindmillTestCase):
-
-    def afterSetUp(self):
-        """Setup for each test
-        """
+    def setUp(self):
         ztc.utils.setupCoreSessions(self.app)
-        self.setRoles(['Manager'])
-        self.login_user()
 
-    def test_exporter(self):
-        client = self.wm
+    # def test_exporter(self):
+    #     client = self.wm
 
-        # Create a new page with the title 'Page'
-        client.open(url=u'/plone')
-        client.waits.forPageLoad(timeout=u'20000')
+    #     # Create a new page with the title 'Page'
+    #     client.open(url=u'/plone')
+    #     client.waits.forPageLoad(timeout=u'20000')
         #client.execJS(js=u"$('.deco-plone\\\\.app\\\\.standardtiles\\\\.' + \
         #                     'title-tile h1').html('Page')")
         #client.mouseDown(jquery=u'(".deco-button-save")[0]')